+++ NEWS.md
@@ -33,6 +33,7 @@
 * `source_gist()` works once more when there is only a single file
   in the gist (#1266).
 * Infrastructure functions (`use_*`) now use the implementations in **usethis**.
+* Use rcmdcheck to run and parse R CMD check output (#1153).
 
 # devtools 1.13.2
 Workaround a regression in Rcpp::compileAttributes.